chore: separate comment component
